requirements.txt
pymongo==4.7.2
motor==3.4.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.4.2
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
jupyter==1.0.0
matplotlib==3.9.0
seaborn==0.13.2